+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Changed
+
+- [TD-3245] Tested compatibility with PostgreSQL 9.6, 10.15, 11.10, 12.5 and
+  13.1. CI pipeline changed to use `postgres:12.5-alpine`.
+  
 ### Added
 
 - [TD-3296] Support for PKCE in OAuth 2.0 authentication code flow. To enable